feat: create emailController file that holds the transporter and sendMail function. modify: move transporter and sendmail function over to controller file
