+++ src/index.js
@@ -21,43 +21,10 @@ dbSetup();
 //==================================================
 // Routes
 //==================================================
-
-//Placeholder routes for webpages
-// app.get('/', (req, res) => {
-//   res.send('Welcome to the Tour Info App');
-// });
 app.use(locationRoutes);
-// app.use(emailRoutes);
+app.use(emailRoutes);
 app.use('/', index);
-//==================================================
-// Nodemailer
-//==================================================
-const transporter = nodemailer.createTransport({
-  port: 465,
-  host: "smtp.gmail.com",
-  auth: {
-    user: 'rickmayatest@gmail.com',
-    pass: 'ChappedLips!234'
-  },
-  secure: true
-});
 
-app.post('/text-mail', (req, res) => {
-  const { to, subject, text } = req.body;
-  const mailData = {
-    from: 'rickmayatest@gmail.com',
-    to: to,
-    subject: subject,
-    text: text
-  };
-
-  transporter.sendMail(mailData, (error, info) => {
-    if (error) {
-      return console.log(error);
-    }
-    res.status(200).send({ message: "Mail sent", message_id: info.messageId });
-  });
-});
 
 //SEEDERS
 const {seedCities} = require('./seeders/citySeeder');
